requirement.txt
Django==3.2.9
djangorestframework==3.12.4
PyJWT==1.7.1
djangorestframework_simplejwt==4.4.0
drf-yasg==1.20.0
Pillow